requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
sqlalchemy==2.0.23
aiosqlite==0.19.0

# GitHub integration
PyGithub==1.59.1

# AI/ML dependencies
openai==1.3.7
numpy==1.24.3
faiss-cpu==1.7.4

# Other utilities
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.32.3
